requirements.txt
Django==5.0.2
django-crispy-forms==2.1
crispy-bootstrap5==2024.2
python-dotenv==1.0.1
requests==2.31.0
beautifulsoup4==4.12.3
nltk==3.8.1
openai==1.12.0
django-allauth==0.61.1
django-cleanup==8.1.0
Pillow==10.2.0
django-taggit==5.0.1
django-filter==23.5
django-rest-framework==0.1.0 